+++ .eslintrc
@@ -6,7 +6,14 @@
       "error", { "capIsNew": false }
     ],
     "import/no-extraneous-dependencies": [
-      "error", { "devDependencies": ["test/**/*.js", "**/*.test.js", "**/*.spec.js"] }
+      "error", {
+        "devDependencies": [
+          "test/**/*.js",
+          "**/*.test.js",
+          "**/*.spec.js",
+          "**/ts-types-tester.js"
+        ]
+      }
     ],
     "comma-dangle": [
       "error", {
