+++ src/targets/amplitude/index.js
@@ -1,3 +1,1 @@
-const { Amplitude } = require('./amplitude');
-
-module.exports = { Amplitude };
+module.exports = require('./amplitude');
