+++ src/targets/google-analytics/__tests__/google-analytics-node-env.test.js
@@ -2,11 +2,11 @@
  * @jest-environment node
  */
 
-const { GoogleTagManager } = require('../../src/targets/google-tag-manager');
+const { GoogleAnalytics } = require('../google-analytics');
 
 describe('If window does not exist', () => {
   it('should just silently not send events', () => {
     const events = [{ hitType: 'pageview' }];
-    expect(() => GoogleTagManager(events)).not.toThrow();
+    expect(() => GoogleAnalytics(events)).not.toThrow();
   });
 });
