+++ src/targets/google-analytics/__tests__/ts-types.js
@@ -1,0 +1,3 @@
+const { tsTypesTester } = require('../../../utils/ts-types-tester');
+
+tsTypesTester(__dirname);
