+++ src/targets/google-tag-manager/__tests__/google-tag-manager-node-env.test.js
@@ -2,13 +2,11 @@
  * @jest-environment node
  */
 
-const {
-  GoogleAnalytics,
-} = require('../../src/targets/google-analytics');
+const { GoogleTagManager } = require('../google-tag-manager');
 
 describe('If window does not exist', () => {
   it('should just silently not send events', () => {
     const events = [{ hitType: 'pageview' }];
-    expect(() => GoogleAnalytics(events)).not.toThrow();
+    expect(() => GoogleTagManager(events)).not.toThrow();
   });
 });
