+++ src/targets/react-native/google-analytics/google-analytics.js
