+++ src/targets/react-native/google-analytics/index.d.ts
