+++ src/targets/react-native/google-tag-manager/google-tag-manager.js
