+++ src/targets/react-native/google-tag-manager/index.d.ts
