--- src/targets/react-native/index.js
@@ -1,7 +1,0 @@
-const { GoogleAnalytics } = require('./google-analytics');
-const { GoogleTagManager } = require('./google-tag-manager');
-
-module.exports = {
-  GoogleAnalytics,
-  GoogleTagManager,
-};
