+++ src/utils/add-timestamp.js
@@ -2,4 +2,6 @@ function addTimestamp(events) {
   return events.map(event => Object.assign({}, event, { timeSaved: Date.now() }));
 }
 
-module.exports = addTimestamp;
+module.exports = {
+  addTimestamp,
+};
